fix: send service worker errors back to caller (#28)
Add `sendError` method to `postmessage` to return errors thrown back to
the caller.

This makes it so that any error thrown in a `postmessage` command will
be automatically returned to the caller as part of the promise. This
makes error handling more simple, and more likely to catch unforeseen
errors. I was able to remove some `try` clauses, because the errors will
be caught and sent back to the main thread automatically now.

I did this work as I was trying to track down a bug where sometimes
`/swfs/init` returns 404, resulting in `sw.id` not being set correctly.
This commit does not actually fix the service worker issue, but at least
raises the error to the user when it happens:


![sw-failed](https://github.com/hrgdavor/jscadui/assets/1766297/7402ee41-c666-49d0-a106-46b6690b461d)