fix other apps (#57)
As we've been evolving the jscadui project with a focus on jscad.app...
some of the other apps got broken along the way. Users are still asking
about how to host standalone model pages, so it would be good to keep
them working.

Most of the breakage was from changing `runFile` to `runScript`.

I also took the liberty of updating the `jscad-examples` app with a new
design... it is a parameterized dropdown that let's you select from the
other examples.


![jscad-examples](https://github.com/hrgdavor/jscadui/assets/1766297/91a4ba53-1b36-4358-8767-76b7fd41e17f)

I also updated the READMEs, mostly fixing typos and adding some more
details.