WIP: feat-v1-support (#102)
Initial implementation of support for V1 scripts.

test url while in progress https://3d.hrg.hr/jscad/v1-support/ or run
locally


V1 is not really supported anymore, this is an effort to allow running
V1 scripts for many
old scripts, but with limited dedication what kind of issues would be
fixed. Certainly is very unlikely
existing bugs from v1 will be fixed (unless someone is very eager and
cotributes a fix)

- [ ] export support (not working yet)
- [x] 2d shapes (outlines, poly)
- [x] colors support (working, but missing default color for partially
colored)
- [ ] complex projects with include (may or may not be implemented)
- [x] drag drop `.jscad` and shim will be added to the script



Bonus over old V1 engine
 - correct line numbers for errors reported
- debugger support matching source (jsut write `debugger` in the script
somewhere and if dev console is open it will brak on that line)


------------------

it works relatively ok, will need a followup. merging to not have it on the side for too long