Allow reload after changes in imported files (#124)
This changes do several things:

1. Fix the cache clearing after modifications in imported files. The
fs-provider uses the file path. The cache uses the URL. I have added
some logic to generate the URL from the path. This URL can then be used
to clear the cache.
2. Implement logic for clearing the cache in indirectly imported files.
The new Map requireCache.knownDependencies contains a list of all files
that are imported for a specific file. This information is used to
recursively clear the caches of other modules when one of its
dependencies has changed.
3. Add more JSDoc Type annotations
4. Some minor syntax cleanups (let => const, || => ??, self=>globalThis,
...)

This updates for changes in the main file are still broken due to a bug
in a prior commit.

---------

Co-authored-by: Davor Hrg <hrgdavor@gmail.com>