Add type information to jscad-web (#128)
I have activated strict ts typechecks in vscode. This resulted in a lot
of warnings. I have used these warnings to add missing type information
to a lot of places in the code. The project is still far away from full
type safety but its (hopefully) a step in the right direction.

I know this PR is pretty large, a bit all over the place and hard to
review. Sorry for that. Please contact me for questions, critique or
anything else.